fusión de rama-juan-v1 con restaurante a tope

+++ backend/database/migrations/20240523030538-create-productingredient.js
@@ -10,7 +10,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       quantity: {
-        type: Sequelize.FLOAT
+        type: Sequelize.DOUBLE
       },
       ingredientId: {
         type: Sequelize.INTEGER,
